epan: Avoid overflow calculating number of bytes required 7 bit encoding
Calculating the number of bytes required for a certain number of 7 bit
characters at a certain bit offset doesn't overflow in the result via
the natural method, but an intermediate result must be stored in a
larger type to avoid overflow. Alternatively, the number of characters
could be split into the quotient and remainder with 8 in order to divide
by 8 first; that is probably slower here on 64-bit architectures with
32-bit inputs and output, but might be worth it on 32-bit architectures
or 64-bit inputs and outputs.

In the signed case, intermediate overflow of an initially positive
signed integer (as expected here) leads to a negative result; division
here would definitely leave it negative whereas the right bit shift is
implementation defined but probably also arithmetic and retains the
sign. This leads to handling as a negative number and in many case either
taking the rest of the tvb or throwing an exception, instead of wrapping
to a smaller number. Hence the same PoC that works on the current master
may not produce an error on released versions. (See #20103.)

Fix #21549, #21550

+++ epan/proto.c
@@ -14359,7 +14359,7 @@ proto_tree_add_ts_23_038_7bits_packed_item(proto_tree *tree, const int hfindex, 
 
 	DISSECTOR_ASSERT_FIELD_TYPE(hfinfo, FT_STRING);
 
-	byte_length = (((no_of_chars + 1) * 7) + (bit_offset & 0x07)) >> 3;
+	byte_length = (int)(((((int64_t)no_of_chars + 1) * 7) + (bit_offset & 0x07)) >> 3);
 	byte_offset = bit_offset >> 3;
 
 	string = tvb_get_ts_23_038_7bits_string_packed(PNODE_POOL(tree), tvb, bit_offset, no_of_chars);
@@ -14387,7 +14387,7 @@ proto_tree_add_ascii_7bits_item(proto_tree *tree, const int hfindex, tvbuff_t *t
 
 	DISSECTOR_ASSERT_FIELD_TYPE(hfinfo, FT_STRING);
 
-	byte_length = (((no_of_chars + 1) * 7) + (bit_offset & 0x07)) >> 3;
+	byte_length = (int)(((((int64_t)no_of_chars + 1) * 7) + (bit_offset & 0x07)) >> 3);
 	byte_offset = bit_offset >> 3;
 
 	string = tvb_get_ascii_7bits_string(PNODE_POOL(tree), tvb, bit_offset, no_of_chars);
